@@ -4,25 +4,27 @@
   "private": true,
   "proxy": "http://localhost:9999",
   "dependencies": {
-    "@postlight/mercury-parser": "^2.2.0",
-    "@types/jest": "^24.9.1",
-    "@types/node": "^12.12.67",
-    "@types/postlight__mercury-parser": "^2.2.3",
-    "@types/react": "^16.9.51",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-router-dom": "^4.3.5",
-    "antd": "^4.0.0",
+    "@emotion/react": "^11.4.1",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.0.3",
+    "@mui/lab": "^5.0.0-alpha.50",
+    "@mui/material": "^5.0.3",
+    "@postlight/mercury-parser": "^2.2.1",
+    "@types/jest": "^27.4.1",
+    "@types/node": "^17.0.23",
+    "@types/postlight__mercury-parser": "^2.2.4",
+    "@types/react-dom": "^17.0.14",
+    "@types/react-router-dom": "^5.3.3",
     "decimal.js-light": "^2.5.1",
-    "lossless-json": "^1.0.4",
-    "moment": "^2.29.1",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-list": "^0.8.15",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "^3.4.4",
+    "lossless-json": "^1.0.5",
+    "moment": "^2.29.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-list": "^0.8.17",
+    "react-router-dom": "^4.0.3",
+    "react-scripts": "5.0.0",
     "react-scroll": "^1.8.1",
-    "simplebar-react": "^2.3.5",
-    "typescript": "^4.1.3"
+    "typescript": "^4.6.3"
   },
   "scripts": {
     "analyze": "source-map-explorer build/static/js/main.*",
@@ -32,11 +34,14 @@
     "eject": "react-scripts eject"
   },
   "devDependencies": {
-    "@types/react-list": "^0.8.5",
-    "@types/react-router": "^5.1.8",
+    "@types/react": "^17.0.43",
+    "@types/react-list": "^0.8.7",
+    "@types/react-router": "^5.1.18",
     "@types/react-scroll": "^1.8.0",
-    "@typescript-eslint/parser": "^4.11.0",
-    "source-map-explorer": "^1.8.0"
+    "@typescript-eslint/parser": "^5.18.0",
+    "eslint": "^7.32.0",
+    "eslint-plugin-react": "^7.29.4",
+    "source-map-explorer": "^2.5.2"
   },
   "browserslist": [
     ">0.2%",
