@@ -1,8 +1,13 @@
-import React, {FormEvent, ReactNode} from 'react';
-import Form from 'antd/lib/form';
-import Input from 'antd/lib/input';
-import Button from 'antd/lib/button';
+import React, {ChangeEvent, FormEvent, ReactNode} from 'react';
 import {RouteComponentProps, withRouter} from "react-router-dom";
+import {
+  Box,
+  Button,
+  createTheme,
+  CssBaseline,
+  TextField,
+  ThemeProvider
+} from "@mui/material";
 
 // WrappedLoginProps needs to extend RouteComponentProps to get "history".
 export interface WrappedLoginProps extends RouteComponentProps {
@@ -11,64 +16,95 @@ export interface WrappedLoginProps extends RouteComponentProps {
 class WrappedLogin extends React.Component<WrappedLoginProps, any> {
   constructor(props: WrappedLoginProps) {
     super(props);
-    this.state = {loginFailed: false};
+    this.state = {
+      loginFailed: false,
+      usernameMissing: false,
+      passwordMissing: false,
+      username: "",
+      password: ""
+    };
   }
 
   render() {
-    return (
-      <div className='login-page'>
-        <div className='login-page-logo'>Goliath</div>
-        <div className='login-page-form-background'>
-          <Form
-            id="login-form"
-            onFinish={(e: FormEvent) => this.handleSubmit(e)}
-            layout='vertical'
-            className='login-page-form'>
-            {this.showLoginFailedMessage()}
+    const theme = createTheme();
 
-            <Form.Item
-              name="username"
-              rules={[{required: true, message: 'Empty username!'}]}>
-              <Input
-                prefix={<i
-                  className="fas fa-user"
-                  aria-hidden="true"/>}
-                placeholder="username"/>
-            </Form.Item>
-            <Form.Item
-              name="password"
-              rules={[{required: true, message: 'Empty password!'}]}>
-              <Input
-                prefix={<i
-                  className="fas fa-lock"
-                  aria-hidden="true"/>}
-                type="password"
-                placeholder="password"/>
-            </Form.Item>
-            <Button
-              type="primary"
-              htmlType="submit"
-              className="login-page-form-button">
-              Log in
-            </Button>
-          </Form>
-        </div>
-      </div>
+    return (
+      <ThemeProvider theme={theme}>
+        <CssBaseline/>
+        <Box className='GoliathLoginPage'>
+          <Box className='GoliathLoginPageLogo'>Goliath</Box>
+          <Box className='GoliathLoginPageSecondary'>
+            <Box
+              className="GoliathLoginPageForm">
+              {this.showLoginFailedMessage()}
+              <form
+                onSubmit={(e: FormEvent) => this.handleSubmit(e)}
+                className="login-page-form">
+                <TextField
+                  name="username"
+                  label="Username"
+                  error={this.state.usernameMissing}
+                  value={this.state.username}
+                  required
+                  helperText={this.state.usernameMissing ? "Missing Username" : ""}
+                  onChange={(e: ChangeEvent) => this.handleChange(e)}
+                />
+                <TextField
+                  name="password"
+                  label="Password"
+                  type="password"
+                  error={this.state.passwordMissing}
+                  value={this.state.password}
+                  required
+                  helperText={this.state.passwordMissing ? "Missing Password" : ""}
+                  onChange={(e: ChangeEvent) => this.handleChange(e)}
+                />
+                <Button
+                  variant="contained"
+                  type="submit"
+                  onClick={(e: FormEvent) => this.handleSubmit(e)}
+                  className="GoliathLoginFormButton">
+                  Log in
+                </Button>
+              </form>
+            </Box>
+          </Box>
+        </Box>
+      </ThemeProvider>
     )
   }
 
   showLoginFailedMessage = (): ReactNode => {
     if (this.state.loginFailed) {
-      return <div className="login-failed">Invalid username or password!</div>
+      return <Box className="GoliathLoginFailedMessage">
+        Invalid username or password.
+      </Box>
     } else {
       return null;
     }
   };
 
+  handleChange(e: any) {
+    this.setState({
+      [e.target.name]: e.target.value
+    })
+  }
+
   handleSubmit(e: React.FormEvent) {
+    e.preventDefault();
+
+    if (!this.validateForm()) {
+      return;
+    }
+
+    const data = {
+      'username': this.state.username,
+      'password': this.state.password
+    }
+
     fetch('/auth', {
       method: 'POST',
-      body: JSON.stringify(e),
+      body: JSON.stringify(data),
       credentials: 'include'
     }).then((res: Response) => {
       if (!res.ok) {
@@ -84,6 +120,20 @@ class WrappedLogin extends React.Component<WrappedLoginProps, any> {
       this.setState({loginFailed: true});
       console.log(e);
     });
+  }
+
+  validateForm() {
+    const usernameMissing = this.state.username === "";
+    const passwordMissing = this.state.password === "";
+
+    if (usernameMissing || passwordMissing) {
+      this.setState({
+        usernameMissing,
+        passwordMissing
+      });
+      return false;
+    }
+    return true;
   }
 }
 
